% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{correlacao}
\alias{correlacao}
\title{Calcula a Correlação entre Variáveis}
\usage{
correlacao(data, dependent_vars, independent_vars, method = "pearson")
}
\arguments{
\item{data}{Data frame contendo os dados.}

\item{dependent_vars}{Vetor com os nomes das variáveis dependentes.}

\item{independent_vars}{Vetor com os nomes das variáveis independentes.}

\item{method}{Método de correlação (padrão é "pearson").}
}
\value{
Matriz de correlações.
}
\description{
Esta função calcula a correlação entre variáveis dependentes e independentes.
}
