% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{lag_comp}
\alias{lag_comp}
\title{Cria Lags de uma Variável}
\usage{
lag_comp(data, lags, var)
}
\arguments{
\item{data}{Data frame contendo os dados.}

\item{lags}{Número de lags a serem criados.}

\item{var}{Nome da variável para a qual os lags serão criados.}
}
\value{
Data frame com as colunas de lags adicionadas.
}
\description{
Esta função cria colunas com lags de uma variável específica.
}
