% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{media_movel}
\alias{media_movel}
\title{Calcula Média Móvel}
\usage{
media_movel(data, var, k)
}
\arguments{
\item{data}{Data frame contendo os dados.}

\item{var}{Nome da variável para a qual a média móvel será calculada.}

\item{k}{Ordem (número de períodos) da média móvel.}
}
\value{
Data frame com a coluna da média móvel adicionada.
}
\description{
Esta função calcula a média móvel de uma variável específica em um data frame.
}
